Replaced Base Project
The project has been replaced. Before, it was the project with the platformer assets from the lecture. Now it is the base project with Bolt installed but no extra assets added. I also added the platformer assets that will be used to make the game, which were downloaded from the Unity Store.